Set up framework for rapid prototyping
The Game class knows about Players, who have Pieces. The Game class
also knows about the View and Scene that contains the Items. Pieces are
ellipses while Squares are rectangular. There is a single square in the
view.

On the GUI side, there are four buttons to add pieces to the square
instance in the scene. Clicking on the buttons should remove them.
There's a functional quit button.